# go/docs/swagger.yaml
definitions:
  controller.MaterialListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Material'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      ts:
        type: string
    type: object
  model.Material:
    properties:
      material_desc:
        type: string
      material_id:
        type: integer
      material_location:
        type: string
      material_name:
        type: string
      material_purchased_at:
        type: string
      material_status:
        type: integer
      material_type_id:
        type: integer
    type: object
  model.MaterialStats:
    properties:
      available_total:
        type: integer
      by_type:
        items:
          $ref: '#/definitions/model.MaterialStatsByType'
        type: array
    type: object
  model.MaterialStatsByType:
    properties:
      available_count:
        type: integer
      material_type_id:
        type: integer
      material_type_name:
        type: string
      total_count:
        type: integer
    type: object
  model.MaterialType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /material-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MaterialType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取材料类型列表
      tags:
      - 材料类型管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 材料类型信息
        in: body
        name: materialType
        required: true
        schema:
          $ref: '#/definitions/model.MaterialType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.MaterialType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建材料类型
      tags:
      - 材料类型管理
  /material-types/{id}:
    delete:
      parameters:
      - description: 材料类型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除材料类型
      tags:
      - 材料类型管理
    get:
      parameters:
      - description: 材料类型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MaterialType'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取材料类型详情
      tags:
      - 材料类型管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 材料类型ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新后的材料类型信息
        in: body
        name: materialType
        required: true
        schema:
          $ref: '#/definitions/model.MaterialType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MaterialType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新材料类型
      tags:
      - 材料类型管理
  /materials:
    get:
      parameters:
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认 20
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.MaterialListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 物资列表
      tags:
      - 物资管理
    post:
      consumes:
      - application/json
      description: 新增一条物资记录
      parameters:
      - description: 物资信息
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建物资
      tags:
      - 物资管理
  /materials/{id}:
    delete:
      parameters:
      - description: 物资ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除物资
      tags:
      - 物资管理
    get:
      parameters:
      - description: 物资ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取物资详情
      tags:
      - 物资管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 物资ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Material'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新物资
      tags:
      - 物资管理
  /materials/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MaterialStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 物资统计
      tags:
      - 物资管理
swagger: "2.0"